DMP-3536 Calculate Confidence for associated object - Identify business Rules
Added tests to calculate retention confidence score and reason for associated objects
